+++ server/src/auth/error.rs
@@ -1,0 +1,6 @@
+#[derive(Debug)]
+pub enum AuthError {
+    FailedToGetJwtValidator,
+    MissingAuthHeader,
+    InvalidToken,
+}
